% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/citations.R
\name{citations}
\alias{citations}
\title{Retrieves most popular articles over a defined time period}
\usage{
citations(day, page = NULL, num_results = 100, cited_in = NULL,
  doi_prefix = NULL, nlmid = NULL, subject = NULL, foptions = list())
}
\arguments{
\item{day}{The time period over which metrics are required. Allowed options include '1d', '2d', '3d', '4d', '5d', '6d', '1w', '1m', '3m', '6m', '1y'resu}

\item{page}{Page number}

\item{num_results}{Max number of results per page. Cannot exceed \code{100}.}

\item{cited_in}{One or more comma delimited options from: facebook, blogs, linkedin, video, pinterest, gplus, twitter, reddit, news, f1000, rh, qna, forum, peerreview}

\item{doi_prefix}{A DOI prefix (the bit before the first slash, e.g. 10.1038)}

\item{nlmid}{Comma delimited list of journal NLM IDs. Include only articles from journals with the supplied NLM journal IDs (only journals indexed in PubMed have NLM IDs).}

\item{subject}{Comma delimited list of slugified journal subjects. Include only articles from journals matching any of the supplied NLM subject ontology term(s).}

\item{foptions}{A list of additional arguments for \code{httr}. There is no reason to use this argument except for debugging purposes.}
}
\description{
The time period can be any of the following: '1d', '2d', '3d', '4d', '5d', '6d', '1w', '1m', '3m', '6m', '1y'
}
\examples{
citations(day = '1d')
# Only Facebook mentions
fb_1week <- citations('1w', cited_in = "facebook")
}

