% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics}
\usage{
metrics(identifier)
}
\arguments{
\item{identifier}{The identifier passed in the format "identifier/id".}
}
\description{
Returns metrics for any standard identifier such as doi, arxiv, pmid, id, or ads.
}
\examples{
\dontrun{
foo <- metrics(identifier = "arxiv/1108.2455")
# This is a failed example
foo <- metrics(identifier = "arxiv/1108.24553")
# Now for a PMID
pm <- metrics(identifier = "pmid/21148220")
}
}

