% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.altmetric}
\alias{plot.altmetric}
\title{Plots metrics for an altmetric object}
\usage{
\method{plot}{altmetric}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Altmetric}}

\item{...}{additional arguments}
}
\description{
Plots metrics for an altmetric object
}

