% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\name{print.altmetric}
\alias{print.altmetric}
\title{Print a summary for an altmetric object}
\usage{
\method{print}{altmetric}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Altmetric}}

\item{...}{additional arguments}
}
\description{
Print a summary for an altmetric object
}
