\name{return_provider}
\alias{return_provider}
\title{Returns cleaner metric source names (mostly for internal use)}
\usage{
  return_provider(provider)
}
\arguments{
  \item{provider}{the data provider}
}
\description{
  Returns cleaner metric source names (mostly for internal
  use)
}
\examples{
\dontrun{
return_provider('cited_by_gplus_count')
}
}
\keyword{internal}

