% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{altmetric_data}
\alias{altmetric_data}
\title{Returns a data.frame from an S3 object of class altmetric}
\usage{
altmetric_data(alt_obj)
}
\arguments{
\item{alt_obj}{An object of class altmetric}
}
\description{
Returns a data.frame from an S3 object of class altmetric
}
