\name{altmetric_data}
\alias{altmetric_data}
\title{Returns a data frame of metrics for a paper}
\usage{
  altmetric_data(alt_obj)
}
\arguments{
  \item{alt_obj}{altmetrics object}
}
\value{
  data.frame
}
\description{
  Returns a data frame of metrics for a paper
}
\examples{
\dontrun{
altmetric_data(altmetrics('10.1038/489201a'))
}
}

