% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{altmetrics}
\alias{altmetrics}
\title{Query data from the Altmetric.com API}
\usage{
altmetrics(oid = NULL, id = NULL, doi = NULL, pmid = NULL,
  arXiv = NULL, apikey = getOption("altmetricKey"), foptions = list(),
  ...)
}
\arguments{
\item{oid}{An object ID}

\item{id}{An altmetric.com id for a scholarly paper}

\item{doi}{A DOI}

\item{pmid}{A pubmed ID}

\item{arXiv}{An Arxiv id.}

\item{apikey}{Your API `key`. By default the package ships with a key, but mostly as a demo. If the key becomes overused, then it is likely that you will start to see API limit errors}

\item{foptions}{Additional options for `httr`}

\item{...}{additional options}
}
\description{
Query data from the Altmetric.com API
}
\examples{
z <- altmetrics(doi ='10.1038/480426a')
# If you desire a `list` instead, then simply turn off the `flatten` argument
}
